day_four/requirements.txt
flask
flask-sqlalchemy
flask-pydantic
pydantic[email]
sqlalchemy
psycopg2-binary
alembic
pytest
pytest-flask
python-dotenv
httpx